Refactors of Audio Visualizer and Web RTC pong
Cleaned up the code and reorganized Web RTC pong to make it simpler and easier to read. Also added status messages and styling.

Audio Visualizer was cleaned up to make the code more consistant, minor variable renames (Audio was renamed to Sound as Audio is also part of the DOM)