Refactor and Added new BG Effect
Made more element id's consistent.
moved data and arr variables to a their proper location in the Sound object and gave them more fitting names of frequency a waveform.
Updated the background/fade shader to display the frequency data, will be revised later to use a texture to import a higher resolution frequency data.
Reduced the frequency data size to support import of the data to the shader.
Added background color smoothing, to prepare for the merge of this with the original visualizer.