Preparing for Cube Solver V3.2
We gave cube solver v.3.1 a last quick re format as we start work on v3.2 which aims to completely refactor and clean up the code as well as improve it quite a bit as it is honestly confusing and in consistent at times. Plans are in the top of the main code, more are to come.