Major background change for Cube Solver
Nearly everything but rendering was changed to be more object oriented. It does run slower than the original and is unable to solve some cubes the original could, there is a lot of room for optimization as well as it will be easier to do so with the more readable code!