Switched to Modules
Biggest Changes:
To help with readability as well as better abstraction and preparation for use in web workers, most classes have been separated into individual modules.
A major bug was discovered in the solving code (imagine sorting a list and then completely ignoring it!) which prevented it from actually solving pretty much any cube that did not start with L' as the solution. Now the previous major changes and bug fixes from the refactor now come through as an improvement over the old version in that it can now solve the super flip in less than 3 hours (now taking less than 2 minutes)

Note: The algorithm still needs to be worked on. It comes up with some interesting solutions to simple cubes (such as the checkerboard pattern).

Planned Work:
- GUI
- Settings/Options
- Play by play of solution
- Lots of improvements
- Mobile Support ( Refactor seams to have degraded it)