Possible optimization added
Code was added to sort cubes into the large list when solving in sorted batches in order to take better use of each sweep of inserting cubes. This seems to have improved some solves but increased the number of cycles and time to do others. This change should not have done that..., I will look into it later, but it works! There is also a lot of invalid state catches that will trigger the debugger added in.